Moved all constants back into data.py.
